Adding --keep-seeds and --generate-functions command line options.

+++ yinyang/src/core/Fuzzer.py
@@ -200,7 +200,8 @@ class Fuzzer:
                     break  # Continue to next seed.
 
                 self.statistic.mutants += 1
-                os.remove(scratchfile)
+                if not self.args.keep_mutants:
+                    os.remove(scratchfile)
 
             log_finished_generations(successful_gens, unsuccessful_gens)
         self.terminate()
